+++ src/types/deprecated.ts
@@ -1,0 +1,78 @@
+import { FileUploadParams } from "./files";
+import {
+  CreateRealtimeTemporaryTokenParams,
+  LemurActionItemsParams,
+  LemurBaseParams,
+  LemurQuestionAnswerParams,
+  LemurSummaryParams,
+  LemurTaskParams,
+  ListTranscriptParams,
+  TranscriptOptionalParams,
+  TranscriptParams,
+} from "./openapi.generated";
+import { SubmitParams, TranscribeParams } from "./transcripts";
+
+/**
+ * @deprecated
+ * Use`FileUploadParams` instead.
+ */
+export type FileUploadParameters = FileUploadParams;
+
+/**
+ * @deprecated
+ * Use`TranscribeParams` instead.
+ */
+export type TranscribeParameters = TranscribeParams;
+
+/**
+ * @deprecated
+ * Use`SubmitParams` instead.
+ */
+export type SubmitParameters = SubmitParams;
+
+/**
+ * @deprecated
+ * Use`CreateRealtimeTemporaryTokenParams` instead.
+ */
+export type CreateRealtimeTemporaryTokenParameters =
+  CreateRealtimeTemporaryTokenParams;
+/**
+ * @deprecated
+ * Use`LemurActionItemsParams` instead.
+ */
+export type LemurActionItemsParameters = LemurActionItemsParams;
+/**
+ * @deprecated
+ * Use`LemurBaseParams` instead.
+ */
+export type LemurBaseParameters = LemurBaseParams;
+/**
+ * @deprecated
+ * Use`LemurQuestionAnswerParams` instead.
+ */
+export type LemurQuestionAnswerParameters = LemurQuestionAnswerParams;
+/**
+ * @deprecated
+ * Use`LemurSummaryParams` instead.
+ */
+export type LemurSummaryParameters = LemurSummaryParams;
+/**
+ * @deprecated
+ * Use`LemurTaskParams` instead.
+ */
+export type LemurTaskParameters = LemurTaskParams;
+/**
+ * @deprecated
+ * Use`ListTranscriptParams` instead.
+ */
+export type TranscriptListParameters = ListTranscriptParams;
+/**
+ * @deprecated
+ * Use`TranscriptOptionalParams` instead.
+ */
+export type CreateTranscriptOptionalParameters = TranscriptOptionalParams;
+/**
+ * @deprecated
+ * Use`TranscriptParams` instead.
+ */
+export type CreateTranscriptParameters = TranscriptParams;