+++ tests/__mocks__/ws.ts
@@ -1,11 +1,11 @@
 import { WebSocket } from "mock-socket";
 
 export default class MockWebSocket extends WebSocket {
-  constructor(address: string | URL, options?: unknown){
+  constructor(address: string | URL, options?: unknown) {
     super(address);
   }
 
-  removeAllListeners(){
+  removeAllListeners() {
     this.listeners = {};
     this.onclose = () => {};
     this.onopen = () => {};